Indeed-Crawler/requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==8.4.0
outcome==1.3.0.post0
packaging==24.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
setuptools==70.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
